requirements.txt
GitPython==2.1.5
nltk==3.2.4

